Refactor to split logic and presentation
